Added functionality to choose button...
Now we can choose a new folder and send its path to render world with electron store. The only problem, I had to restart the app to see the new folder's pics. To prevent this, I added a restart functionality to OK button. It worked for now but this is not exactly what I wanna do. I think it is a good big step on our custom-settings-dialog branch! PS: I got help from AI...